Move API tests over to integration section
